Add factory info support

+++ driver-avalon7.c
@@ -760,6 +760,10 @@ static int decode_pkg(struct cgpu_info *avalon7, struct avalon7_ret *ar, int mod
 			info->get_asic[modular_id][x_miner_id][x_asic_id][10] = tmp;
 		}
 		break;
+	case AVA7_P_STATUS_FAC:
+		applog(LOG_DEBUG, "%s-%d-%d: AVA7_P_STATUS_FAC", avalon7->drv->name, avalon7->device_id, modular_id);
+		info->factory_info[0] = ar->data[0];
+		break;
 	default:
 		applog(LOG_DEBUG, "%s-%d-%d: Unknown response %x", avalon7->drv->name, avalon7->device_id, modular_id, ar->type);
 		break;
@@ -1840,6 +1844,24 @@ static void avalon7_set_freq(struct cgpu_info *avalon7, int addr, int miner_id, 
 		avalon7_iic_xfer_pkg(avalon7, addr, &send_pkg, NULL);
 }
 
+static void avalon7_set_factory_info(struct cgpu_info *avalon7, int addr, uint8_t value[])
+{
+	struct avalon7_pkg send_pkg;
+	uint8_t i;
+
+	memset(send_pkg.data, 0, AVA7_P_DATA_LEN);
+
+	for (i = 0; i < AVA7_DEFAULT_FACTORY_INFO_CNT; i++)
+	      send_pkg.data[i] = value[i];
+
+	/* Package the data */
+	avalon7_init_pkg(&send_pkg, AVA7_P_SET_FAC, 1, 1);
+	if (addr == AVA7_MODULE_BROADCAST)
+		avalon7_send_bc_pkgs(avalon7, &send_pkg);
+	else
+		avalon7_iic_xfer_pkg(avalon7, addr, &send_pkg, NULL);
+}
+
 static void avalon7_set_ss_param(struct cgpu_info *avalon7, int addr)
 {
 	struct avalon7_pkg send_pkg;
@@ -2270,6 +2292,9 @@ static struct api_data *avalon7_api_stats(struct cgpu_info *avalon7)
 		strcat(statbuf, buf);
 
 		if (opt_debug) {
+			sprintf(buf, " FAC0[%d]", info->factory_info[0]);
+			strcat(statbuf, buf);
+
 			for (j = 0; j < info->miner_count[i]; j++) {
 				sprintf(buf, " SF%d[", j);
 				strcat(statbuf, buf);
@@ -2523,6 +2548,30 @@ char *set_avalon7_device_freq(struct cgpu_info *avalon7, char *arg)
 	return NULL;
 }
 
+char *set_avalon7_factory_info(struct cgpu_info *avalon7, char *arg)
+{
+	struct avalon7_info *info = avalon7->device_data;
+	int val;
+
+	if (!(*arg))
+		return NULL;
+
+	sscanf(arg, "%d", &val);
+	if (!val)
+		val = AVA7_DEFAULT_FACTORY_INFO_0;
+
+	if (val < AVA7_DEFAULT_FACTORY_INFO_0_MIN || val > AVA7_DEFAULT_FACTORY_INFO_0_MAX)
+		return "Invalid value passed to set_avalon7_factory_info";
+
+	info->factory_info[0] = val;
+	avalon7_set_factory_info(avalon7, 0, (uint8_t *)info->factory_info);
+
+	applog(LOG_NOTICE, "%s-%d: Update factory info %d",
+		avalon7->drv->name, avalon7->device_id, val);
+
+	return NULL;
+}
+
 static char *avalon7_set_device(struct cgpu_info *avalon7, char *option, char *setting, char *replybuf)
 {
 	unsigned int val;
@@ -2625,6 +2674,15 @@ static char *avalon7_set_device(struct cgpu_info *avalon7, char *option, char *s
 		}
 
 		return set_avalon7_device_voltage(avalon7, setting);
+	}
+
+	if (strcasecmp(option, "factory") == 0) {
+		if (!setting || !*setting) {
+			sprintf(replybuf, "missing factory info");
+			return replybuf;
+		}
+
+		return set_avalon7_factory_info(avalon7, setting);
 	}
 
 	if (strcasecmp(option, "reboot") == 0) {
